PyGame, Tkinter, and Python Chart of Some Colors by Name
Displays a chart of 139 PyGame, Tkinter, and Python colors by name and in some order. Enables the user to see and retrieve names of colors.

Note: I like and use prefixes, and do recognize that I will go straight to hell because of it. Please forgive me, and change as needed.